Introducing a trait for key exchange
